+++ apis/BaseApi.ts
@@ -14,6 +14,7 @@ import { ClientConfig } from '../interfaces/ClientConfig';
 import { tokenHasExpired } from '../utils/Token';
 import { Guid } from '@Commerce-commercetools/utils/Guid';
 import { ExternalError } from '@Commerce-commercetools/errors/ExternalError';
+import { ConfigurationError } from '@Commerce-commercetools/errors/ConfigurationError';
 
 const defaultCurrency = 'USD';
 
@@ -462,13 +463,16 @@ export abstract class BaseApi {
     }
   }
 
-  setSessionCheckoutSessionToken(cartId: string, token: Token): void {
+  async setSessionCheckoutSessionToken(cartId: string, token: Token, locale: Locale): Promise<void> {
+    const checkoutHashKey = await this.getCheckoutHashKey(cartId, locale);
+
     this.sessionData.checkoutSessionToken = {};
-    this.sessionData.checkoutSessionToken[this.getCheckoutHashKey(cartId)] = token;
+    this.sessionData.checkoutSessionToken[checkoutHashKey] = token;
   }
 
-  getSessionCheckoutSessionToken(cartId: string): Token | undefined {
-    return this.sessionData?.checkoutSessionToken?.[this.getCheckoutHashKey(cartId)] ?? undefined;
+  async getSessionCheckoutSessionToken(cartId: string, locale: Locale): Promise<Token | undefined> {
+    const checkoutHashKey = await this.getCheckoutHashKey(cartId, locale);
+    return this.sessionData?.checkoutSessionToken?.[checkoutHashKey] ?? undefined;
   }
 
   protected getAnonymousIdFromSessionData(): string {
@@ -577,8 +581,8 @@ export abstract class BaseApi {
     return project;
   }
 
-  protected async generateCheckoutSessionToken(cartId: string) {
-    const checkoutSessionToken = this.getSessionCheckoutSessionToken(cartId);
+  protected async generateCheckoutSessionToken(cartId: string, locale: Locale) {
+    const checkoutSessionToken = await this.getSessionCheckoutSessionToken(cartId, locale);
 
     if (!tokenHasExpired(checkoutSessionToken)) {
       // The token exist and is not expired, so we don't need to generate a new one.
@@ -587,7 +591,7 @@ export abstract class BaseApi {
 
     if (checkoutSessionToken) {
       try {
-        return await this.refreshCheckoutSessionToken(cartId, checkoutSessionToken);
+        return await this.refreshCheckoutSessionToken(cartId, checkoutSessionToken, locale);
       } catch (error) {
         // We are ignoring the error refreshing the token and trying to generate a new one
       }
@@ -602,14 +606,14 @@ export abstract class BaseApi {
         },
       },
       metadata: {
-        applicationKey: this.clientSettings.checkoutApplicationKey,
+        applicationKey: this.getCheckoutApplicationKey(locale),
       },
     });
 
-    return await this.fetchCheckoutSessionToken(cartId, url, body);
+    return await this.fetchCheckoutSessionToken(cartId, url, body, locale);
   }
 
-  private async refreshCheckoutSessionToken(cartId: string, checkoutSessionToken: Token) {
+  private async refreshCheckoutSessionToken(cartId: string, checkoutSessionToken: Token, locale: Locale) {
     const url = `${this.clientSettings.sessionUrl}/${this.projectKey}/sessions/${checkoutSessionToken.token}`;
 
     const body = JSON.stringify({
@@ -620,10 +624,10 @@ export abstract class BaseApi {
       ],
     });
 
-    return await this.fetchCheckoutSessionToken(cartId, url, body);
+    return await this.fetchCheckoutSessionToken(cartId, url, body, locale);
   }
 
-  private async fetchCheckoutSessionToken(cartId: string, url: string, body: string) {
+  private async fetchCheckoutSessionToken(cartId: string, url: string, body: string, locale: Locale) {
     const headers = {
       'Content-Type': 'application/json',
       Authorization: `Bearer ${this.token.token}`,
@@ -657,7 +661,7 @@ export abstract class BaseApi {
       expirationTime: response?.expiryAt ? new Date(response?.expiryAt).getTime() : undefined,
     };
 
-    this.setSessionCheckoutSessionToken(cartId, token);
+    await this.setSessionCheckoutSessionToken(cartId, token, locale);
 
     return token;
   }
@@ -691,14 +695,16 @@ export abstract class BaseApi {
     })();
   }
 
-  private getCheckoutHashKey(cartId: string): string {
+  private async getCheckoutHashKey(cartId: string, locale: Locale): Promise<string> {
     if (this.checkoutHashKey) {
       return this.checkoutHashKey;
     }
 
+    const checkoutApplicationKey = this.getCheckoutApplicationKey(locale);
+
     this.checkoutHashKey = crypto
       .createHash('md5')
-      .update(this.clientSettings.checkoutApplicationKey + cartId)
+      .update(checkoutApplicationKey + cartId)
       .digest('hex');
 
     return this.checkoutHashKey;
@@ -738,5 +744,17 @@ export abstract class BaseApi {
     this.apiRoot = createApiBuilderFromCtpClient(client);
 
     return this.apiRoot;
+  }
+
+  private getCheckoutApplicationKey(locale: Locale): string {
+    const checkoutApplicationKey = this.clientSettings.checkoutApplicationKey[locale.country];
+
+    if (!checkoutApplicationKey) {
+      throw new ConfigurationError({
+        message: `There is no checkout application key defined for country ${locale.country} `,
+      });
+    }
+
+    return checkoutApplicationKey;
   }
 }